Wifi: Add WCNSS_QTI_AOSP CFLAG for value-added AOSP
This commit moves the addition of WCNSS_QTI_AOSP flag
from the make file of qcwcn to the make file of Wifi_hal.
This flag was earlier cleared while building wifi_hal
when present in qcwcn's Android.mk .

Thus, this change adds this required flag in the corresponding
make file[wifi_hal's Android.mk] so that it is used in wifi_hal
during compilation.

Change-Id: Ie4a131ed2a6e0080479e20a60b05b7c4024ca1d0
CRs-Fixed: 2329296
Signed-off-by: unsatifsed27 <unsatifsederror@gmail.com>